# 8262896: [macos_aarch64] Crash in jni_fast_GetLongField