--- a/test/hotspot/gtest/runtime/test_threads.cpp
+++ b/test/hotspot/gtest/runtime/test_threads.cpp
@@ -176,3 +176,16 @@ TEST_VM(ThreadsTest, claim_overflow) {
   ThreadInVMfromNative invm(JavaThread::current());
   VMThread::execute(&op);
 }
+
+TEST_VM(ThreadsTest, fast_jni_in_vm) {
+  JavaThread* current = JavaThread::current();
+  JNIEnv* env = current->jni_environment();
+  MACOS_AARCH64_ONLY(ThreadWXEnable wx(WXWrite, current));
+
+  // DirectByteBuffer is an easy way to trigger GetIntField,
+  // see JDK-8262896
+  jlong capacity = 0x10000;
+  jobject buffer = env->NewDirectByteBuffer(NULL, (jlong)capacity);
+  ASSERT_NE((void*)NULL, buffer);
+  ASSERT_EQ(capacity, env->GetDirectBufferCapacity(buffer));
+}