Hi, please review a fix for a random crash on macos/aarch64.

By default, GetXXXField JNI Interface implementation is a generated function (-XX:+UseFastJNIAccessors). Usually the function is called by JNI code running in WXExec mode and everything is fine. But sometime we attempt to call it in WXWrite context, like in the stack trace attached to the bug:

```
v  ~BufferBlob::jni_fast_GetLongField
V  [libjvm.dylib+0x7a6538]  Perf_Detach+0x168
j  jdk.internal.perf.Perf.detach(Ljava/nio/ByteBuffer;)V+0 java.base@17-internal
j  jdk.internal.perf.Perf$CleanerAction.run()V+8 java.base@17-internal
j  jdk.internal.ref.CleanerImpl$PhantomCleanableRef.performCleanup()V+4 java.base@17-internal
j  jdk.internal.ref.PhantomCleanable.clean()V+12 java.base@17-internal
j  jdk.internal.ref.CleanerImpl.run()V+57 java.base@17-internal
j  java.lang.Thread.run()V+11 java.base@17-internal
j  jdk.internal.misc.InnocuousThread.run()V+20 java.base@17-internal
v  ~StubRoutines::call_stub
```

One way to fix the bug is to ensure WXExec mode before calling GetXXXField, but it depends on finding and fixing all such cases. 

This patch instead adds additional actions to GetXXXField implementation to ensure correct W^X mode regardless if it is called from WXWrite or WXExec mode.

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue
- [x] Change must be properly reviewed

### Issue
 * [JDK-8262896](https://bugs.openjdk.java.net/browse/JDK-8262896): [macos_aarch64] Crash in jni_fast_GetLongField


### Reviewers
 * [Andrew Haley](https://openjdk.java.net/census#aph) (@theRealAph - **Reviewer**)
 * [David Holmes](https://openjdk.java.net/census#dholmes) (@dholmes-ora - **Reviewer**)


### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.java.net/jdk pull/3422/head:pull/3422` \
`$ git checkout pull/3422`

Update a local copy of the PR: \
`$ git checkout pull/3422` \
`$ git pull https://git.openjdk.java.net/jdk pull/3422/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 3422`

View PR using the GUI difftool: \
`$ git pr show -t 3422`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.java.net/jdk/pull/3422.diff">https://git.openjdk.java.net/jdk/pull/3422.diff</a>

</details>