@theRealAph 2021-04-10
So I'm curious because I don't know the overhead of W^X switching. How much does this gain over turning off UseFastJNIAccessors? I guess it doesn't much matter because you keep track of the current mode and only flip it if you really need to, which is rare?