@theRealAph 2021-04-13
Looking at this further, the use of W^X for functions like this one, callable from JNI code, makes no sense.
These functions are never written to once they have been generated. Could we generate them during init into a non-executable page, then remap that page as read/execute only? Then we wouldn't have to care about the state of W^X.