@theRealAph 2021-04-13
> It depends on performance. I've had a quick look at how `pthread_jit_write_protect_np` works, and it looks like some serious overhead. Can you measure, say, a few million JNI accessors to see which works best?

I just had a look, and according to [](https://dougallj.github.io/applecpu/firestorm.html) MSR (APRR) only takes a single cycle, but there is still some overhead because of an `ISB`. Please try it and see, though.