@VladimirKempik 2021-04-13
> Why not? As far as I can see, the "Allow Unsigned Executable Memory" Hardened Runtime capability allows exactly this. Maybe it'd stop OpenJDK being notarized, but Apple doen't say so.

AFAIR, this entitlement https://developer.apple.com/documentation/bundleresources/entitlements/com_apple_security_cs_allow-unsigned-executable-memory is not available on macos_arm64

this entitlement allows you to allocate rwx memory on macos_intel, but on macos_arm64 kernel forbids to allocate rwx memory(withotu MAP_JIT) no matter what.