@AntonKozlov 2021-04-13
Thanks! Yes, I also not happy with this from performance side. I'll look in this.

/integrate